+++ contracts/Interfaces/ICurrency.sol
@@ -18,5 +18,5 @@ interface ICurrency {
     /**
      * @dev burn token for recipient. Assuming onlyGovernance
      */
-    function burn(address account, uint amount) external;
+    function burn(address account, uint256 amount) external;
 }
